+++ cruise-control/src/test/java/com/linkedin/kafka/cruisecontrol/KafkaCruiseControlUnitTestUtils.java
@@ -25,6 +25,9 @@ import static com.linkedin.kafka.cruisecontrol.config.KafkaCruiseControlConfig.D
 public class KafkaCruiseControlUnitTestUtils {
   public static final int ANOMALY_DETECTOR_INITIAL_QUEUE_SIZE = 10;
   public static final String OPENAPI_SPEC_PATH = System.getProperty("user.dir") + "/src/yaml/base.yaml";
+  public static final String CRUISE_CONTROL_PACKAGE = "com.linkedin.kafka.cruisecontrol";
+  public static final String JSON_CONTENT_TYPE = "application/json";
+  public static final String PLAIN_TEXT_CONTENT_TYPE = "text/plain";
 
   private KafkaCruiseControlUnitTestUtils() {
